[BCK] web폴더에서 기능별로 views.py 나누기
