+++ web/project/config/production.py
@@ -1,11 +1,11 @@
 from flask_sqlalchemy import SQLAlchemy
 
 db = {
-    'user'      : 'dogma',
-    'password'  : 'Dogma7789!',
-    'host'      : '54.180.131.194',
+    'user'      : 'sixdogma',
+    'password'  : 'Poiu0987*',
+    'host'      : '13.113.12.130',
     'port'      : '3306',
-    'database'  : 'auth'
+    'database'  : 'Flask'
 }
 
 SQLALCHEMY_DATABASE_URI = f"mysql+mysqlconnector://{db['user']}:{db['password']}@{db['host']}:{db['port']}/{db['database']}?charset=utf8"
